N°3793 - Cleanup of orphan CMDBChange can hang the setup

+++ setup/applicationinstaller.class.inc.php
@@ -727,10 +727,15 @@ class ApplicationInstaller
 				SetupPage::log_info("There are $iOrphanCount useless records in {$sDBPrefix}priv_change (".sprintf('%.2f', ((100.0*$iOrphanCount)/$iTotalCount))."%)");
 				if ($iOrphanCount > 0)
 				{
-					SetupPage::log_info("Removing the orphan records...");
-					$sCleanup = "DELETE FROM `{$sDBPrefix}priv_change` USING `{$sDBPrefix}priv_change` LEFT JOIN `{$sDBPrefix}priv_changeop` ON `{$sDBPrefix}priv_change`.id = `{$sDBPrefix}priv_changeop`.changeid WHERE `{$sDBPrefix}priv_changeop`.id IS NULL;";
-					CMDBSource::Query($sCleanup);
-					SetupPage::log_info("Cleanup completed successfully.");
+					if ($iOrphanCount > 100000)
+					{
+						SetupPage::log_warning("There are too much useless records ($iOrphanCount) in {$sDBPrefix}priv_change. Cleanup cannot be done during setup.");
+					} else {
+						SetupPage::log_info("Removing the orphan records...");
+						$sCleanup = "DELETE FROM `{$sDBPrefix}priv_change` USING `{$sDBPrefix}priv_change` LEFT JOIN `{$sDBPrefix}priv_changeop` ON `{$sDBPrefix}priv_change`.id = `{$sDBPrefix}priv_changeop`.changeid WHERE `{$sDBPrefix}priv_changeop`.id IS NULL;";
+						CMDBSource::Query($sCleanup);
+						SetupPage::log_info("Cleanup completed successfully.");
+					}
 				}
 				else
 				{
